refactor: MoveRule 공통된 메서드 분리
