+++ src/main/java/model/piece/moveRule/ChariotMoveRule.java
@@ -1,23 +1,12 @@
 package model.piece.moveRule;
 
 import model.piece.position.Position;
-
-import java.util.ArrayList;
-import java.util.List;
-
 import static model.piece.position.MovementChecker.*;
 
-class ChariotMoveRule implements MoveRule {
+class ChariotMoveRule extends MultiStepMoveRule {
 
     @Override
-    public final List<Position> calculateRouteToDestination(final Position startPosition, final Position destination) {
-        if (canPieceMove(startPosition, destination)) {
-            return findRouteStepByStepToDestination(startPosition, destination);
-        }
-        throw new IllegalArgumentException("현재 기물이 이동할 수 없는 위치입니다.");
-    }
-
-    private boolean canPieceMove(final Position startPosition, final Position destination) {
+    protected boolean canPieceMove(final Position startPosition, final Position destination) {
         int rowStep = destination.calculateRowDifference(startPosition);
         int columnStep = destination.calculateColumnDifference(startPosition);
         if (startPosition.isInPalace() && destination.isInPalace()) {
@@ -41,19 +30,4 @@ class ChariotMoveRule implements MoveRule {
                 || isDownStraight(rowStep, columnStep);
     }
 
-    private List<Position> findRouteStepByStepToDestination(final Position startPosition, final Position destination) {
-        int rowStep = destination.calculateRowDifference(startPosition);
-        int columnStep = destination.calculateColumnDifference(startPosition);
-        List<Position> route = new ArrayList<>();
-        while (isStepRemain(rowStep, columnStep)) {
-            route.add(startPosition.moveIfDestinationIsValid(rowStep, columnStep));
-            rowStep = AbsoluteValueDecreaser.decreaseOne(rowStep);
-            columnStep = AbsoluteValueDecreaser.decreaseOne(columnStep);
-        }
-        return route;
-    }
-
-    private boolean isStepRemain(final int rowStep, final int columnStep) {
-        return rowStep != 0 || columnStep != 0;
-    }
 }
