+++ src/main/java/model/piece/moveRule/JolMoveRule.java
@@ -2,21 +2,12 @@ package model.piece.moveRule;
 
 import model.piece.position.Position;
 
-import java.util.List;
-
 import static model.piece.position.MovementChecker.*;
 
-class JolMoveRule implements MoveRule {
+class JolMoveRule extends OneStepMoveRule {
 
     @Override
-    public final List<Position> calculateRouteToDestination(final Position startPosition, final Position destination) {
-        if (canPieceMove(startPosition, destination)) {
-            return findRouteToDestination(startPosition, destination);
-        }
-        throw new IllegalArgumentException("현재 기물이 이동할 수 없는 위치입니다.");
-    }
-
-    private boolean canPieceMove(final Position startPosition, final Position destination) {
+    protected boolean canPieceMove(final Position startPosition, final Position destination) {
         int rowStep = destination.calculateRowDifference(startPosition);
         int columnStep = destination.calculateColumnDifference(startPosition);
         if (startPosition.isInPalace() && destination.isInPalace()) {
@@ -37,9 +28,4 @@ class JolMoveRule implements MoveRule {
                 || isRight(rowStep, columnStep);
     }
 
-    private List<Position> findRouteToDestination(final Position startPosition, final Position destination) {
-        int rowStep = destination.calculateRowDifference(startPosition);
-        int columnStep = destination.calculateColumnDifference(startPosition);
-        return List.of(startPosition.moveIfDestinationIsValid(rowStep, columnStep));
-    }
 }
